WI #2686 Implement tests outside LSP for GetDataLayout
