+++ TypeCobol.LanguageServer.Test/ProcessorTests/DataLayout/CSV-simple.txt
@@ -1,0 +1,23 @@
+LineNumber;NodeLevel;LevelNumber;VariableName;PictureTypeOrUsage;Start;End;Length
+12;0;1;FILLER (1);GROUP OCCURS 1;1;81;81
+13;1;5;Array1 (1 1);GROUP OCCURS 9;1;81;81
+14;2;10;Var1 (1 1);PIC X;1;1;1
+15;2;10;Array2 (1 1 1);GROUP OCCURS 8;2;9;8
+16;3;15;Var2 (1 1 1);PIC X;2;2;1
+17;0;1;FILLER;GROUP;1;4;4
+18;1;5;GRP-REDEFINED;GROUP;1;3;3
+19;2;10;VAR-1;PIC 9(05) COMP-3;1;3;3
+20;1;5;FILLER;REDEFINES GRP-REDEFINED;1;3;3
+21;2;10;VAR-RED1;PIC 9(05) COMP-3;1;3;3
+22;1;5;GROUP-WITH-88-1;GROUP;4;4;1
+23;2;10;GROUP-WITH-88-2;PIC 9(01);4;4;1
+28;0;1;Var1-in-ls;PIC XX;1;2;2
+29;0;1;num1;PIC 9(03);1;3;3
+30;0;1;num2;comp-1;1;4;4
+31;0;1;num3;comp-2;1;8;8
+32;0;1;num4;PIC 9(03) comp-3;1;2;2
+33;0;1;num5;PIC 9(03) comp-4;1;2;2
+34;0;1;num6;PIC 9(03) comp-5;1;2;2
+35;0;1;num7;PIC 9(03) packed-decimal;1;2;2
+36;0;1;num8;PIC 9(03) binary;1;2;2
+37;0;1;num9;PIC 9(03) comp;1;2;2
