+++ TypeCobol.LanguageServer.Test/ProcessorTests/DataLayout/CSV-stackedAndNested.txt
@@ -1,0 +1,3 @@
+LineNumber;NodeLevel;LevelNumber;VariableName;PictureTypeOrUsage;Start;End;Length
+6;0;1;main-group;GROUP;1;1;1
+7;1;5;main-var;PIC X;1;1;1
